refactor(mercury): split encryption.rs under 700-line cap; PR-review fixes
Address PR-review findings on the v2 encryption foundation:

- Strip #NNN issue refs from source comments (encryption + Cargo.toml);
  provenance lives in the PR, not tracked source.
- Fix the v2 IV rustdoc: the IV comes from a CSPRNG (`rand`, seeded from
  the OS), not an OS CSPRNG syscall per packet.
- Convert the flat encryption.rs (~985 lines, over the 700-line hard cap)
  to a directory module: impl stays in encryption/mod.rs, the test block
  moves verbatim into encryption/tests.rs (`#[cfg(test)] mod tests;`).
  Pure move + the two comment fixes — no impl or test-logic changes.

Both resulting files are under 700 lines (mod.rs 502, tests.rs 482).
All 242 cimmeria-mercury tests still pass.
