+++ crates/mercury/Cargo.toml
@@ -28,7 +28,7 @@ async-trait = { workspace = true }
 
 # Cryptography
 #  - v1 (legacy, byte-identical to the C++ client): AES-256-CBC + HMAC-MD5.
-#  - v2 (#434 Phase 3): AES-256-CBC + HMAC-SHA256 (truncated to 16 bytes),
+#  - v2: AES-256-CBC + HMAC-SHA256 (truncated to 16 bytes),
 #    per-packet random IV, HKDF-SHA256 key separation.
 aes = { workspace = true }
 cbc = { workspace = true }
